Move /analysis/ to /internal/analysis/
