+++ internal/analysis/analysis.go
